requirements.txt
gradio==4.37.2
pypdf==4.2.0
pyyaml==6.0.1

langchain==0.3.7
langchain-core==0.3.12
langchain-community==0.3.3
langchain-openai==0.2.3

tqdm==4.66.4
